When INSERTing, use NULL when parent doesn't have specified field
This is "adresseerbaarobjectnevenadres" has 3 fields, only one of which
will be present in the _parent object (ligplaatsstatus, standplaatsstatus
or verblijfplaatsstatus).

Because the database uses enum types for these values, and those enums
don't contain the empty string, inserting these records breaks.

By inserting NULLs instead, everything works the same as when using the
"COPY" statement.